% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{dffm.criterion}
\alias{dffm.criterion}
\title{dffm criterion}
\usage{
dffm.criterion(x, K.max = NULL, p.max = NULL)
}
\arguments{
\item{x}{An object which has to be of class 'dffm'.}

\item{K.max}{A predetermined maximum of K for the criterion.}

\item{p.max}{A predetermined maximum of p for the criterion.}
}
\value{
An optimal number of K and p based on the information criteria from Otto and Salish (2021) and all possible combinations of K
and p stored in a matrix.
}
\description{
Computes an optimal number of factors K and an optimal number of lags p for functional data analysis or prediction, based on the
information criterion by Otto and Salish (2021).
}
\examples{
data = load.dieboldli()
d = dffm(d)
dffm.criterion(d)
}
\references{
\itemize{
\item Otto S. and Salish N. (2021). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
}
}
