% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.forecast.R
\name{dffm.forecast}
\alias{dffm.forecast}
\title{dffm.forecast}
\usage{
dffm.forecast(dffmobj, h = 1)
}
\arguments{
\item{dffmobj}{An object of class 'dffm'.}

\item{h}{A parameter determining the forecast horizon. If h is NULL, h will be 1. NULL is default.}
}
\value{
A list containing the following:
\item{predfactors}{A vector containing the predicted factors for further computations.}
\item{predcurves.obsgrid}{A matrix or ts-matrix containing the predicted curves evaluated on the observationgrid.}
\item{predcurves.workgrid}{A matrix or ts-matrix containing the predicted curves evaluated on the workinggrid.}
}
\description{
A function that computes the MSE-optimal h-step adead curve predictor for a given 'dffm' object.
}
\examples{
fed = load.fed()
d = dffm(fed, K = 3, p = 2, AR = FALSE)
dffm.forecast(d, h = 2)
}
