% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.MSEplot.R
\name{dffm.MSEplot}
\alias{dffm.MSEplot}
\title{Plot of the MSE of different combinations of factors and lags}
\usage{
dffm.MSEplot(criterion, kmax, pmax, rotate = NULL, cex.main = NULL, ...)
}
\arguments{
\item{criterion}{Must be an output of the fts.criterion function.}

\item{kmax}{A parameter specifying the maximum number of factors used in the 3D plot.}

\item{pmax}{A parameter specifying the maximum number lags used in the 3D plot.}

\item{rotate}{Parameter to change the horizontal rotation of the 3D plot. If NULL rotate will be set to 120.
NULL is default.}

\item{cex.main}{Parameter to change the font size of the headline. If NULL cex.main will be set to 1.2.
NULL is default.}

\item{...}{Option to change parameters in surf3D().}
}
\value{
A 3D plot of the mse errors from the dffm.criterion function.
}
\description{
A function that produces a 3D-plot of the mean squared errors from the dffm.criterion function. For more information about the
mse, see the fts.criterion function.
}
\examples{
# with dffm.criterion
JKV = load.JKV()
JKV.fda = fda.preprocess(data = JKV)
dffm.MSEplot(fts.criterion(JKV.fda, p.max = 8), kmax = 15, pmax = 8)
}
