% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{dffm.predict}
\alias{dffm.predict}
\title{dffm predict}
\usage{
dffm.predict(x, AR = FALSE, criterion = FALSE, h = NULL, p = NULL, K = NULL)
}
\arguments{
\item{x}{An object which has to be of class 'dffm'.}

\item{AR}{Will determine if prediction is based on vector autoregressive models or autoregressive models. If AR is FALSE
prediction will be based on vector autoregressive models. FALSE is default.}

\item{criterion}{Automatic selection of the number of factors K and lags p. If TRUE the consistent information criterion of Otto and Salish (2021)
is used. If FALSE the value specified in 'K' and 'p' is used. FALSE is default. For more details see 'dffm.criterion'.}

\item{h}{Selects the amount of predictions which will be made. If h is NULL, h will be 1. NULL is default.}

\item{p}{Selects the lags which will be used in the prediction. If p is NULL, p will be 8. NULL is default. If criterion is
TRUE p will be overwritten.}

\item{K}{Selects the number of factors which will be used for the prediction. If K is NULL, K will be maximum of possible
components. NULL is default. If criterion is TRUE K will be overwritten.}
}
\value{
\item{predicted.values}{A matrix or a ts-matrix containing the predicted fitted data.}
\item{predicted.factors}{A matrix or a ts-matrix containing the predicted factors.}
}
\description{
Will make predictions of the fitted input data from a 'dffm' object based on vector autoregressive or autoregressive models.
}
\examples{
data = load.dieboldli()
d = dffm(d)
dffm.predict(d)
}
\references{
\itemize{
\item Otto S. and Salish N. (2021). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
}
}
