% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{dffm.preprocessing}
\alias{dffm.preprocessing}
\title{dffm preprocessing}
\usage{
dffm.preprocessing(
  data,
  pace = FALSE,
  observationgrid = NULL,
  user.gridsize = NULL,
  lambda = NULL
)
}
\arguments{
\item{data}{A multivariate time series of class 'ts' or 'data.frame'.}

\item{pace}{Preprocessing method to be used. If TRUE the PACE method from the package 'fdapace' is applied (see Yao et al., 2005).
In case of irregularly observed functional data please set to TRUE.
If FALSE smoothing splines from the package 'fda' are used (see Kokoszka and Reimherr, 2017). FALSE is default.}

\item{observationgrid}{An optional list, sequence or numeric vector which contains a predetermined observationgrid for the
observations. If NULL the column names will be used. NULL is default.}

\item{user.gridsize}{An optional parameter specifying the size of the workinggrid. If user.gridsize is NULL, user.gridsize will be the highest
number from the column names or the specified observationgrid. NULL is default.}

\item{lambda}{A smoothing parameter for the penalty term which prevents over fitting. If lambda is 0, the penalty term will
not contribute to the analysis. Lambda is only used in the 'FUNCTIONAL DATA ANALYSIS' ('fd') method. If lambda is NULL,
lambda will be 10^{-3}. NULL is default.}
}
\value{
An object of class 'fd' if pace is FALSE, otherwise an object of class 'FPCA'.
}
\description{
dffm preprocessing is used to prepare a time series or data frame data object especially for the dffm function. For more details
see dffm. The output objects are based on the fda package ('fd') or on the fdapace package ('FPCA').
}
\examples{
# standard case
dl = load.dieboldli()
pp = dffm.preprocessing(data = dl, pace = FALSE, observationgrid = NULL, lambda = 10^{-4})
x = dffm(pp, K = 3)
# pace method
pp = dffm.preprocessing(data = dl, pace = TRUE, user.gridsize = 180)
x = dffm(pp)
}
\references{
\itemize{
\item Kokoszka P. and Reimherr M. (2017), "Introduction to Functional Data Analysis". CRC Press.
\item Yao, F., Mueller, H.-G., and Wang, J.-L. (2005). Functional data analysis for sparse longitudinal
data. Journal of the American Statistical Association, 100:577-590.
}
}
