% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{dffm}
\alias{dffm}
\title{dffm}
\usage{
dffm(
  X,
  pace = FALSE,
  criterion = FALSE,
  observationgrid = NULL,
  user.gridsize = NULL,
  K = NULL
)
}
\arguments{
\item{X}{A multivariate time series of class 'ts' or 'data.frame', an object of the class 'fd' from the package 'fda',
or an object of the class 'FPCA' from the package 'fdapace'.}

\item{pace}{Preprocessing method to be used. If TRUE the PACE method from the package 'fdapace' is applied (see Yao et al., 2005).
In case of irregularly observed functional data please set to TRUE.
If FALSE smoothing splines from the package 'fda' are used (see Kokoszka and Reimherr, 2017). FALSE is default.}

\item{criterion}{Automatic selection of the number of factors K. If TRUE the consistent information criterion of Otto and Salish (2021)
is used. If FALSE the value specified in 'K' is used. FALSE is default. For more details see 'dffm.criterion'.}

\item{observationgrid}{An optional list, sequence or numeric vector which contains a predetermined observationgrid for the
observations. If NULL the column names or input observationgrid from the 'fd' or 'FCPA' object will be used.
NULL is default.}

\item{user.gridsize}{An optional parameter specifying the size of the workinggrid. If user.gridsize is NULL, user.gridsize will be the highest
number from the column names or the specified observationgrid. NULL is default.}

\item{K}{A parameter which specifies the number of factors be used. If K is NULL, K will be 3. NULL is default.}
}
\value{
dffm will return an object of class 'dffm' with:
\item{basis}{A list containing the call of the function; input observationgrid; used workinggrid; if cirterion is TRUE, the
optimal number of lags for prediction p; the number of used components K and input data if it was a data frame or ts object.}
\item{eigenvalues}{A vector containing the first K eigenvalues of the sample covariance operator.}
\item{loadingfunctions}{A matrix containing the estimated loading functions, which are evaluated on the workinggrid.}
\item{meanfunction}{A vector containing the estimated mean function, which is evalauted on the workinggrid.}
\item{factorscores}{A matrix/ts-matrix of the computed factorscores.}
\item{fitted.data}{A matrix/ts-matrix of the fitted curves evaluated on the workinggrid.}
}
\description{
dffm is a function which accompanies the fully functional factor model method from Otto and Salish (2021).
Furthermore does dffm perform the the pace method for sparse data from Yao et al. (2005) if needed.
}
\examples{
# standard case
dl = load.dieboldli()
x = dffm(X = dl, pace = FALSE, criterion = FALSE, observationgrid = NULL, user.gridsize = NULL, K = 3)
x
summary(x)
plot(x, xlab = "maturities", ylab = "yields", date.on = TRUE, time = 20)
dffm.criterion(x = x) # optimal K = 3 and p = 1
p = predict(x = x, AR = FALSE, criterion = FALSE, h = 10, p = 1, K = 3)
plot(y = p$predicted.values[3,], x = 1:120, type = "l", xlab = "maturities", ylab = "yields", main = "predicted yields mar 2001")
# using dffm.preprocessing
pp = dffm.preprocessing(data = dl, pace = FALSE, lambda = 10^{-6})
x = dffm(X = pp, criterion = TRUE)
x$basis$p.opt  # optimal p = 1
x$basis$K  # optimal K = 3
predict(x, p = 1, K = 3, h = 3)
}
\references{
\itemize{
\item Kokoszka P. and Reimherr M. (2017), "Introduction to Functional Data Analysis". CRC Press.
\item Otto S. and Salish N. (2021). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
\item Yao, F., Mueller, H.-G., and Wang, J.-L. (2005). Functional data analysis for sparse longitudinal
data. Journal of the American Statistical Association, 100:577-590.
}
}
\seealso{
\itemize{
\item summary: summary of the dffm object.
\item print: print of the dffm object.
\item predict: predict of the dffm object. For more details, see dffm.predict.
\item plot: plot of the dffm object at specified time with or without date.
\item fitted: output of the fitted values of the dffm object.
}
}
