% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.VARforecast.R
\name{DNS.forecast}
\alias{DNS.forecast}
\title{Dynamic Nelson-Siegel Forecasting for Yield Curves}
\usage{
DNS.forecast(
  fdaobj,
  p = 1,
  AR = FALSE,
  obsdata = FALSE,
  start = NULL,
  end = NULL,
  h = 1
)
}
\arguments{
\item{fdaobj}{An object of class \code{fdaobj}.}

\item{p}{Integer defining the number of lags to include in the VAR model.}

\item{AR}{A logical flag indicating whether to model individual AR(p) processes (TRUE) for each factor, or a collective VAR(p) model (FALSE). Default is FALSE.}

\item{obsdata}{A logical flag indicating whether to use the raw data (TRUE) or the preprocessed data (FALSE) from \code{fdaobj}. Default is FALSE.}

\item{start}{Optional integer specifying the start index for a restricted sample period.}

\item{end}{Optional integer specifying the end index for a restricted sample period.}

\item{h}{An integer specifying the forecast horizon. If h=0, an in-sample fit is returned. For h > 0, an out-of-sample h-step forecast is generated. Default is h=1.}
}
\value{
The function returns different outputs based on the value of h.
If h > 0: Returns the h-step ahead forecasted curve.
If h = 0: Returns a list containing:
\item{curve.predict}{The in-sample predicted curves.}
\item{factors.predict}{The in-sample predicted factors.}
\item{NSloadings}{The Nelson Siegel loadings.}
\item{betas}{The estimated betas (factors) in the Nelson-Siegel model.}
\item{VARmatrix}{The estimated VAR matrix.}
}
\description{
\code{DNS.forecast} implements the Dynamic Nelson-Siegel (DNS) model of Diebold and Li (2006) to forecast yield curves. The function accommodates both Vector Autoregression (VAR) and Autoregressive (AR) models. This function is designed to work with preprocessed functional data objects (\code{fdaobj}).
}
\examples{
# Load, preprocess data, and perform in-sample prediction
fed = load.fed()
fdaobj = fda.preprocess(data = fed)
in_sample_fit = DNS.forecast(fdaobj, h=0)

# Perform 1-step ahead prediction
one_step_ahead = DNS.forecast(fdaobj, h=1)
}
\references{
\itemize{
\item Diebold, F. X., & Li, C. (2006). Forecasting the term structure of government bond yields. Journal of Econometrics, 130(2), 337-364.
}
}
