% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.criterion.R
\name{fts.criterion}
\alias{fts.criterion}
\title{Information Criteria to Estimate the Number of Factors for Functional Factor Models}
\usage{
fts.criterion(fdaobj, K.max = NULL, p.max = NULL)
}
\arguments{
\item{fdaobj}{An object of class 'fdaobj', typically the output from \code{fda.preprocess} or \code{fts.cumAC}.}

\item{K.max}{An optional integer specifying the maximum number of factors to consider. If NULL (default), \code{K.max} will be set to the highest number supported by the given 'fdaobj'.}

\item{p.max}{An optional integer specifying the maximum number of lags to consider. If NULL (default), \code{p.max} will be determined based on Schwert's rule of thumb.}
}
\value{
Returns a list containing:
\item{IC.min}{A data.frame with the estimated number of factors (K) and lags (p) based on the Bayesian Information Criterion (BIC), Hannan-Quinn Criterion (HQC), and the final prediction error (fFPE) criterion of Aue et al. (2015).}
\item{MSE.matrix}{A matrix of Mean Squared Error (MSE) values for all combinations of considered factors and lags.}
}
\description{
\code{fts.criterion} estimates the optimal number of factors (K) and lags (p) for functional factor models based on information criteria. It applies the methodology proposed by Otto and Salish (2024)
}
\details{
Estimates the number of factors K and lags p for the approximate functional factor model using the
information criteria proposed in Otto and Salish (2024).
}
\examples{
fed = load.fed()
fpcaobj = fda.preprocess(data = fed)
criterion_output = fts.criterion(fpcaobj)
print(criterion_output$IC.min)
}
\references{
\itemize{
\item Otto, S., & Salish, N. (2024). Approximate Factor Models For Functional Time Series. arXiv:2201.02532.
\item Aue, A., Norinho, D. D., & Hörmann, S. (2015). On the prediction of stationary functional time series. Journal of the American Statistical Association, 110(509), 378-392.
}
}
