% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.cumAC.R
\name{fts.cumAC}
\alias{fts.cumAC}
\title{Cumulative Outer Product Autocovariance Eigendecomposition for Functional Time Series}
\usage{
fts.cumAC(fdaobj, q0 = 1, start = NULL, end = NULL)
}
\arguments{
\item{fdaobj}{An object of class 'fdaobj', typically the output from \code{fda.preprocess}, representing preprocessed functional time series data.}

\item{q0}{The number of cumulative lags to consider in the autocovariance computation. Default is 1.}

\item{start}{An optional integer specifying the start index for the period over which the eigendecomposition is calculated. If not provided, analysis starts from the beginning of the data series.}

\item{end}{An optional integer specifying the end index for the period over which the eigendecomposition is calculated. If not provided, analysis goes until the end of the data series.}
}
\value{
Returns an object of class 'fdaobj', containing the following components:
\item{densedata}{The dense data from the input object.}
\item{workinggrid}{The wirkinggrod from the input object.}
\item{operator}{The operator for which the eigenelements are computed. Here: "cumAC_operator".}
\item{scores}{The coefficients representing the projections of 'densedata' onto each eigenfunction.}
\item{eigenfunctions}{A matrix of orthonormal eigenfunctions derived from the cumulative outer product autocovariance operator.}
\item{eigenvalues}{The eigenvalues associated with each eigenfunction, indicating the autocovariance explained by each eigenfunction.}
\item{scores.centered}{The projection coefficients after demeaning 'densedata'.}
\item{meanfunction}{The sample mean function from the input object.}
\item{raw.data}{The raw data from the input object.}
\item{observationgrid}{The observation grid from the input object.}
}
\description{
This function performs the eigendecomposition of the cumulative outer product autocovariance operator, denoted D, for functional time series. It is used to estimate factors and loadings in the context of the Otto and Salish (2024) approximate factor model.
}
\examples{
# Example with default settings on a standard working grid
fed = load.fed()
fdaobj = fda.preprocess(data = fed)
fts.cumAC(fdaobj)
}
