% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{load.dieboldli}
\alias{load.dieboldli}
\title{load dieboldli}
\usage{
load.dieboldli()
}
\value{
A time series of yield curve data with maturities ranging from 1 to 120 month.
}
\description{
A function which takes yield curve data provided by Diebold and Li (2006): https://www.sas.upenn.edu/~fdiebold/papers/paper49/FBFITTED.txt
}
\examples{
data = load.dieboldli()
}
\references{
\itemize{
\item Diebold, F. X. and Li, C. (2006). Forecasting the term structure of government bond yields. Journal of Econometrics, 130:337-364.
}
}
