% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.JKV.R
\name{load.JKV}
\alias{load.JKV}
\title{load jungbacker}
\usage{
load.JKV()
}
\value{
Time series of yield curve data containing maturities ranging from 3 to 120 month, which had
been observed from 1987 to 2008.
}
\description{
A function which takes yield curve data provided by Jungbacker et.al (2014) from the
following website: http://qed.econ.queensu.ca/jae/2014-v29.1/jungbacker-koopman-van_der_wel/
}
\examples{
data = load.JKV()
}
\references{
\itemize{
\item Jungbacker, B., Koopman, S. J., and Van der Wel, M. (2014). Smooth dynamic factor analysis with application to the US term structure of
interest rates. Journal of Applied Econometrics, 29:65-90.
}
}
