% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{load.jungbacker}
\alias{load.jungbacker}
\title{load jungbacker}
\usage{
load.jungbacker()
}
\value{
A time series of yield curve data with maturities ranging from 3 to 120 month.
}
\description{
A function which takes yield curve data provided by Jungbacker et.al (2014): http://qed.econ.queensu.ca/jae/2014-v29.1/jungbacker-koopman-van_der_wel/
}
\examples{
data = load.jungbacker()
}
\references{
\itemize{
\item Jungbacker, B., Koopman, S. J., and Van der Wel, M. (2014). Smooth dynamic factor analysis with application to the US term structure of
interest rates. Journal of Applied Econometrics, 29:65-90.
}
}
