% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.LW.R
\name{load.LW}
\alias{load.LW}
\title{load Liu-Wu Yield Data}
\usage{
load.LW()
}
\value{
Annualized continuously-compounded zero-coupon yields in percentage points. Each column corresponds to a maturity between 1 to 360 months.
}
\description{
Function which provides data from Liu-Wu Yield Data via the following website:
https://sites.google.com/view/jingcynthiawu/yield-data
}
\details{
Yan Liu and Jing Cynthia Wu "Reconstructing the Yield Curve", Journal of Financial Economics, 2021, 142 (3), 1395-1425.
}
\examples{
data = load.LW()
}
