% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{plot.dffm}
\alias{plot.dffm}
\title{plot.dffm}
\usage{
\method{plot}{dffm}(x, time = NULL, date.on = FALSE, workgrid = TRUE, ...)
}
\arguments{
\item{x}{An object of class "dffm".}

\item{time}{Observation to be plotted.}

\item{date.on}{Logical; if TRUE, date will be shown underneath plot. If FALSE, no date will be shown. FALSE is default.}

\item{workgrid}{Logical; If TRUE, workgrid will be used. If FALSE observationgrid will be used. TRUE is default.}

\item{...}{Plot adjustment parameters left open.}
}
\description{
S3 predict function for dffm objects.
}
