% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{predict.dffm}
\alias{predict.dffm}
\title{predict.dffm}
\usage{
\method{predict}{dffm}(x, h = NULL)
}
\arguments{
\item{x}{An object of class "dffm".}

\item{h}{Time ahead forecasts.}
}
\value{

}
\description{
S3 predict function for dffm objects.
}
