% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{print.dffm}
\alias{print.dffm}
\title{print dffm}
\usage{
\method{print}{dffm}(x)
}
\arguments{
\item{x}{An object which has to be of class 'dffm'.}
}
\value{
A comprehension, which shows the class and dimensions of the list objects, stored in the 'dffm' object.
}
\description{
The generic S3 method print for an object of class 'dffm'.
}
\examples{
data = load.dieboldli()
d = dffm(data)
print(d)
# or
d
}
