% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{summary.dffm}
\alias{summary.dffm}
\title{summary dffm}
\usage{
\method{summary}{dffm}(x)
}
\arguments{
\item{x}{An object which has to be of class 'dffm'.}
}
\value{
Will print 'eigenvalues', 'call', 'observationgrid', 'workinggrid', 'K', and 'p.opt' of the 'dffm' object.
}
\description{
The generic S3 method summary for an object of class 'dffm'.
}
\examples{
data = load.dieboldli()
d = dffm(data)
summary(d)
}
