% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.3Dplot.R
\name{dffm.3Dplot}
\alias{dffm.3Dplot}
\title{dffm.3Dplot}
\usage{
dffm.3Dplot(
  fpcaobj,
  domainlab = NULL,
  outputlab = NULL,
  rotate = NULL,
  cex.main = NULL,
  ...
)
}
\arguments{
\item{fpcaobj}{Has to be an object of class 'FPCAobj' or 'dffm'.}

\item{domainlab}{ylab name of the plot.}

\item{outputlab}{zlab name of the plot.}

\item{rotate}{Parameter to change the horizontal rotation of the 3D plot. If NULL rotate will be set to 35.
NULL is default.}

\item{cex.main}{Parameter to change the font size of the headline. If NULL cex.main will be set to 1.6.
NULL is default.}

\item{...}{Option to change parameters in surf3D().}
}
\value{
3D plot of data used in the functions 'fpca.preprocess' or 'dffm'.
}
\description{
A function which 3D-plots the data used in the functions 'fpca.preprocess' or 'dffm'. For more details
look up fpca.preprocess() or dffm().
}
\examples{
JKV = load.JKV()
d = dffm(JKV)
d$FPCA$raw.data
dffm.3Dplot(d)
}
