% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.criterion.R
\name{dffm.criterion}
\alias{dffm.criterion}
\title{dffm.criterion}
\usage{
dffm.criterion(fpcaobj, K.max = NULL, p.max = NULL)
}
\arguments{
\item{fpcaobj}{An object of class 'FPCAobj' or 'dffm'.}

\item{K.max}{A predetermined maximum of K for the criterion. If NULL K.max will be highest possible K from given 'FPCAobj'.
NULL is default.}

\item{p.max}{A predetermined maximum of p for the criterion. If NULL p.max will be base on Schwert rule of thumb. NULL is default.}
}
\value{
A list containing the following:
\item{IC.min}{Computed optimal number of K and p, based on the BIC, HQC and fFPE methods.}
\item{MSE.matrix}{A matrix containing all computed mse combinations.}
}
\description{
Computes an optimal number of factors K and an optimal number of lags p for functional data analysis or prediction, based on the
information criterions written in Otto and Salish (2021).
}
\examples{
# with object of class 'FPCAobj'
fed = load.fed()
fpca = fpca.preprocess(data = fed, method = "splines")
dffm.criterion(fpca)
# with object of class 'dffm'
fed = load.fed()
d = dffm(fed)
dffm.criterion(d)
}
\references{
\itemize{
\item Otto S. and Salish N. (2022). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
}
}
