% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.criterion.R
\name{dffm.criterion}
\alias{dffm.criterion}
\title{dffm.criterion}
\usage{
dffm.criterion(fpcaobj, K.max = NULL, p.max = NULL)
}
\arguments{
\item{fpcaobj}{An object of class 'FPCAobj' or 'dffm'.}

\item{K.max}{The maximum number of factors for the criterion. If NULL K.max will be highest possible number from the given 'FPCAobj'.
NULL is default.}

\item{p.max}{The maximum number of lags for the criterion. If NULL p.max will be based on Schwert's rule of thumb. NULL is default.}
}
\value{
A list containing the following:
\item{IC.min}{Estimated number of factors K and lags p, based on the BIC, HQC and fFPE criterion.}
\item{MSE.matrix}{A matrix containing all computed MSE combinations.}
}
\description{
Estimates the number of factors K and lags p for the functional factor model using the
information criteria proposed in Otto and Salish (2021).
}
\examples{
# using object of class 'FPCAobj'
fed = load.fed()
fpca = fpca.preprocess(data = fed)
dffm.criterion(fpca)
# using object of class 'dffm'
d = dffm(fed)
dffm.criterion(d)
}
\references{
\itemize{
\item Otto S. and Salish N. (2022). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
}
}
