% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.MSEplot.R
\name{dffm.MSEplot}
\alias{dffm.MSEplot}
\title{dffm.MSEplot}
\usage{
dffm.MSEplot(criterion, kmax, pmax, ...)
}
\arguments{
\item{criterion}{Has to be an object of class 'dffm', 'FPCAobj' or results of the dffm.criterion function.}

\item{kmax}{A parameter specifying the factors used in the 3D plot.}

\item{pmax}{A parameter specifying the lags used in the 3D plot.}

\item{...}{Option to change parameters in surf3D().}
}
\value{
A 3D plot of the mse errors from the dffm.criterion function.
}
\description{
A function which plots mean squared errors from the dffm.criterion in 3D. For more information about the
mse errors, see the dffm.criterion function.
}
\examples{
# with dffm.criterion
JKV = load.JKV()
fpca = fpca.preprocess(data = JKV, method = "splines")
dffm.MSEplot(dffm.criterion(fpca, p.max = 8), kmax = 15, pmax = 8)
# with FPCAobj
fpca = fpca.preprocess(data = JKV, method = "splines")
dffm.MSEplot(fpca, kmax = 12, pmax = 10)
}
