% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.preprocess.R
\name{fpca.preprocess}
\alias{fpca.preprocess}
\title{fpca.preprocess}
\usage{
fpca.preprocess(
  data,
  workinggrid = NULL,
  observationgrid = NULL,
  method = c("splines", "pace")
)
}
\arguments{
\item{data}{A multivariate time series of class 'ts' or 'data.frame'.}

\item{workinggrid}{An optional list, sequence or numeric vector which contains a predetermined workinggrid for the analysis.
If NULL a sequence, beginning from the smallest to the highest observed observation will be used. NULL is default.}

\item{observationgrid}{An optional list, sequence or numeric vector which contains a predetermined observationgrid for the
observations. If NULL the column names will be used. NULL is default.}

\item{method}{Preprocessing method to be used. If 'splines' smoothing splines from the package 'fda' are used (see Kokoszka
and Reimherr, 2017). If 'pace' the PACE method from the package 'fdapace' is applied (see Yao et al., 2005). In case of
irregularly observed functional data please set to 'pace'. 'splines' is default.}
}
\value{
fpca.preprocess will return an object of class 'FPCAobj' with:
\item{eigenvalues}{A vector containing the eigenvalues of the sample covariance operator.}
\item{fpcascores}{A matrix/ts-matrix of the computed factorscores.}
\item{eigenfunctions.obsgrid}{A matrix containing the estimated loading functions, which are evaluated on the workinggrid.}
\item{meanfunction.obsgrid}{A vector containing the estimated mean function, which is evalauted on the observationgrid.}
\item{eigenfunctions.workgrid}{A matrix containing the estimated loading functions, which are evaluated on the observationgrid.}
\item{meanfunction.workgrid}{A vector containing the estimated mean function, which is evalauted on the workinggrid.}
\item{raw.data}{Ouput of the data used in the analysis.}
\item{observationgrid}{Used observationgrid in the analysis.}
\item{workinggrid}{Used workinggrid in the analysis.}
\item{method}{Used method in the analysis.}
}
\description{
fpca.preprocess is used to prepare a time series or data frame data object especially for the dffm function. For more details
see dffm. The output objects are based on the fda package or on the fdapace package.
}
\examples{
# normal workinggrid
fed = load.fed()
fpca.preprocess(data = fed, method = "splines")
# changed workinggrid
fed = load.fed()
wg = (2:720)/2
fpca.preprocess(fed, workinggrid = wg, method = "splines")
}
\references{
\itemize{
\item Kokoszka P. and Reimherr M. (2017), "Introduction to Functional Data Analysis". CRC Press.
\item Otto S. and Salish N. (2022). "Dynamic Factor Model for Functional Time Series: Identification, Estimation, and Prediction". ...
\item Yao, F., Mueller, H.-G., and Wang, J.-L. (2005). Functional data analysis for sparse longitudinal
data. Journal of the American Statistical Association, 100:577-590.
}
}
