% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{plot.dffm}
\alias{plot.dffm}
\title{plot dffm}
\usage{
\method{plot}{dffm}(x, time = NULL, date.on = FALSE, ...)
}
\arguments{
\item{x}{An object which has to be of class 'dffm'.}

\item{time}{A parameter which selects the wanted fitted data observation. If time is NULL, time will be 1. NULL is default}

\item{date.on}{Will blend in the date of the observation, if it was from a time series object. If date.on is FALSE
the plot will not show the date. FALSE is default.}

\item{...}{Further plot inputs, such as 'ylab' or 'main' are possible.}
}
\value{
A plot of the fitted data.
}
\description{
The generic S3 method plot for an object of class 'dffm'.
}
\examples{
data = load.dieboldli()
d = dffm(data)
plot(d, time = 20, date.on = TRUE)
# or
plot(d, time = 5, xlab = "maturities", ylab = "yields", main = "yield curve of mai 1970")
}
