% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{predict.dffm}
\alias{predict.dffm}
\title{predict.dffm}
\usage{
\method{predict}{dffm}(object, ..., h = NULL)
}
\arguments{
\item{object}{An object of class "dffm".}

\item{...}{further parameters.}

\item{h}{Time ahead forecasts.}
}
\description{
S3 predict function for dffm objects.
}
