% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffm.R
\name{print.dffm}
\alias{print.dffm}
\title{print.dffm}
\usage{
\method{print}{dffm}(x, ...)
}
\arguments{
\item{x}{An object of class "dffm".}

\item{...}{further parameters.}
}
\description{
S3 print function for dffm objects.
}
